feat(repo): refactored admin repo
