+++ backend/domain/use_cases/admin_usecases.py
@@ -1,7 +1,7 @@
 from domain.entities.email import Email
 from domain.entities.admin import Admin
 from domain.value_objects.password import Password
-from domain.repositories.admin_repository import AdminDbRepository
+from tools.for_db.work_with_admin_info import add_admin
 
 
 class AdminRegister:
@@ -15,4 +15,4 @@ class AdminRegister:
         checked_password = Password(self.password)
         hashed_password = checked_password.get_hashed_password()
         admin = Admin(checked_email.get_value(), hashed_password)
-        AdminDbRepository().add_admin(admin)
+        add_admin(admin)
