+++ backend/tests/test_admin_useceses.py
@@ -2,7 +2,7 @@ import pytest
 
 import db.models as models
 from domain.use_cases.admin_usecases import AdminRegister
-from domain.repositories.admin_repository import AdminDbRepository
+from tools.for_db.work_with_admin_info import get_psw_from_db
 from tools.create_db_for_tests import create_test_app_with_db
 from tools.func_for_psw import check_psw
 
@@ -19,6 +19,6 @@ def test_admin_register(app_for_test):
     email = 'mail@.com'
     password = 'Password'
     AdminRegister(email, password).admin_register()
-    psw_from_db = AdminDbRepository().get_psw_from_db(email)
+    psw_from_db = get_psw_from_db(email)
 
     assert check_psw(password, psw_from_db)
